svr: factor out server config
